--- keyboards/bang/default/jef/config.h
@@ -1,66 +1,0 @@
-#pragma once
-// place overrides here
-
-#define COMBO_COUNT 6
-#define TAPPING_TERM 168
-#define COMBO_TERM 160
-
-#define LEADER_PER_KEY_TIMING
-#define LEADER_TIMEOUT 250
-
-// COMBOS!!!!!!
-// Left Hand Combos
-// qw 
-// w+esc 
-// q+esc
-// qd qf
-// vf
-// sx
-
-//Right hand combos
-// '+;
-// jm 
-
-// two hand combos
-// fj qp qk qj vm vk vj 
-
-// Combo Possibilities
-// bq bz 
-// cf* cj* cv cx
-// fq fv* fx fz 
-// hx hz
-// gq gv gx
-// jb jd* jf* jg jh jl jm* jp jq jr js jt jv jw jx jy jz
-// kq kx kz
-// mx mz
-// pq pv px 
-// qb qc qd qf qg qh qj qk ql qm qn qp qq qv qw* qx qy qz
-// sx 
-// tq 
-// vb vf vh vj vk vm vp vq vw vx 
-// wv wx
-// xd xj xk xr xz 
-// yq yy 
-// zf zr zx*
-
-// Special keybinds
-
-// LCTL(kc) || C(kc)                  Hold Left Control + press kc
-// RCTL(kc)                           Hold Right Control + press kc
-// RSFT(kc)                           Hold Right Shift + press kc
-// LSFT(kc) || S(kc)                  Hold Left Shift + press kc
-// LALT(kc) || A(kc)    || LOPT(kc)	  Hold Left Alt + press kc
-// RALT(kc) || ROPT(kc) || ALGR(kc)   Hold Right Alt + press kc
-// LGUI(kc) || G(kc)    || LCMD(kc) || LWIN(kc)   Hold Left GUI + press kc
-// RGUI(kc) || RCMD(kc) || LWIN(kc)   Hold Right GUI + press kc
-
-// SGUI(kc) || SCMD(kc) || SWIN(kc)   Hold Shift + GUI + press kc
-// LCA(kc)                            Hold Control + Alt + press kc
-// LCAG(kc)                           Hold Control, Alt + GUI + press kc
-// MEH(kc)                            Hold Control, Shift + Alt + press kc
-// HYPR(kc)                           Hold Control, Shift, Alt + GUI + press kc
-
-// KC_MEH                                    Left Control, Shift and Alt
-// KC_HYPR                                   Left Control, Shift, Alt and GUI
-
-
