--- keyboards/bang/default/via/rules.mk
@@ -1,3 +1,0 @@
-# MCU name
-MCU = STM32F303
-VIA_ENABLE = yes
