--- keyboards/bang/keymaps/c1x4/info.json
@@ -1,17 +1,0 @@
-{
-    "name": "Dilettante",
-    "vendorId": "0xFEED",
-    "productId": "0x1985",
-    "lighting": "none",
-    "matrix": { "rows": 5, "cols": 15 },
-    "layouts": {
-      "keymap": [
-    [{"x":1},"0,0", "0,1", "0,2", "0,3","0,4","0,5","0,6","0,7","0,8","0,9","0,10","0,11","0,12",{"w":1.5},"0,13"],
-    [{"x":-0.2,"y":-0.5},"1,0", {"x":0.2, "y":0.5,"w":1.5},"1,1", "1,2", "1,3", "1,4", "1,5", "1,6", "1,7", "1,8", "1,9", "1,10", "1,11", "1,12", "1,13", "1,14"],
-    [{"x":-0.2,"y":-0.5},"2,0", {"x":0.2, "y":0.5,"w":1.75},"2,1", "2,2", "2,3", "2,4", "2,5", "2,6", "2,7", "2,8", "2,9", "2,10", "2,11", "2,12", {"w":1.75},"2,13"],
-    [{"x":-0.2,"y":-0.5},"3,0", {"x":0.2, "y":0.5},"3,1",{"w":1.25}, "3,2", "3,3", "3,4", "3,5", "3,6", "3,7", "3,8", "3,9", "3,10", "3,11", "3,12",{"x":0.12,"y":0.2}, "3,13",{"x":0.12, "y":-0.2}, "3,14"],
-    [{"x":-0.2,"y":-0.5},"4,0", {"x":0.2, "y":0.5,"w": 1.5},"4,1",{"w":1.25},"4,2","4,3",{"w": 6},"4,4", "4,5",{"x":0.25, "y":0.2}, "4,6", "4,7", "4,8"]
-    ]
-  }
-}
-
