--- keyboards/bang/keymaps/d1x3/keymap.c
@@ -1,35 +1,0 @@
-#include QMK_KEYBOARD_H
-
-#define TRANS _______
-
-// Defines names for use in layer keycodes and the keymap
-
-// Defines the keycodes used by our macros in process_record_user
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-    [0] = LAYOUT_D_1x3(
-             KC_ESC,  KC_1,    KC_2,  KC_3,  KC_4,  KC_5,  KC_6,  KC_7,  KC_8,   KC_9,    KC_0,     KC_MINS, KC_EQL,  KC_BSPC,
-    KC_VOLU, KC_TAB,  KC_Q,    KC_W,  KC_E,  KC_R,  KC_T,  KC_Y,  KC_U,  KC_I,   KC_O,    KC_P,     KC_LBRC, KC_RBRC, KC_BSLS,
-    KC_VOLD, KC_LCTL, KC_A,    KC_S,  KC_D,  KC_F,  KC_G,  KC_H,  KC_J,  KC_K,   KC_L,    KC_SCLN,  KC_QUOT,          KC_ENT,
-    KC_MUTE, KC_LALT, KC_LSFT, KC_Z,    KC_X,  KC_C,  KC_V,  KC_B,  KC_N,  KC_M,  KC_COMM, KC_DOT, KC_SLSH,  KC_UP,   KC_CAPS,
-             KC_LCTL, KC_LALT, KC_LGUI,               KC_SPACE, KC_SPACE,                   KC_RALT, KC_LEFT, KC_DOWN, KC_RGHT
-    ),
-
-    [1] = LAYOUT_D_1x3(
-             _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,                   _______, _______,
-             _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
-   ),
-
-    [2] = LAYOUT_D_1x3(
-             _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,                   _______, _______,
-             _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
-    )
-
-};
