+++ keyboards/bang/keymaps/default/keymap.c
@@ -1,95 +1,57 @@
 #include QMK_KEYBOARD_H
+#include DEFINE_C
 
 #define TRANS _______
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
-    [0] = LAYOUT_A_1x4(
-          KC_ESC,  KC_1,    KC_2,  KC_3,  KC_4,  KC_5,  KC_6,  KC_7,  KC_8,   KC_9,   KC_0,     KC_MINS, KC_EQL,  KC_BSPC,
-    KC_0, KC_TAB,  KC_Q,    KC_W,  KC_E,  KC_R,  KC_T,  KC_Y,  KC_U,  KC_I,   KC_O,   KC_P,     KC_LBRC, KC_RBRC, KC_BSLS,
-    KC_0, KC_LCTL, KC_A,    KC_S,  KC_D,  KC_F,  KC_G,  KC_H,  KC_J,  KC_K,   KC_L,   KC_SCLN,  KC_QUOT,          KC_ENT,
-    KC_0, KC_LSFT, KC_Z,    KC_X,  KC_C,  KC_V,  KC_B,  KC_N,  KC_M,  KC_COMM, KC_DOT, KC_SLSH,          KC_UP,   KC_CAPS,
-    KC_0, KC_LCTL, KC_LALT, KC_LGUI,                  KC_SPACE,                        KC_RALT, KC_LEFT, KC_DOWN, KC_RGHT
+    [0] = LAYOUT (
+           KC_ESC,  KC_1,    KC_2,  KC_3,  KC_4,  KC_5,  KC_6,  KC_7,  KC_8,    KC_9,   KC_0,    KC_MINS, KC_EQL,  KC_BSPC,
+    PGUP,  KC_TAB,  KC_Q,    KC_W,  KC_E,  KC_R,  KC_T,  KC_Y,  KC_U,  KC_I,    KC_O,   KC_P,    KC_LBRC, KC_RBRC, KC_BSLS,
+    PGDN,  FNCTN,   KC_A,    KC_S,  KC_D,  KC_F,  KC_G,  KC_H,  KC_J,  KC_K,    KC_L,   KC_SCLN, KC_QUOT,          KC_ENT,
+           KC_LSFT, KC_Z,    KC_X,  KC_C,  KC_V,  KC_B,  KC_N,  KC_M,  KC_COMM, KC_DOT, KC_SLSH,          KC_UP,   KC_CAPS,
+    ENCDR, CTRL, KC_LALT, KC_LGUI,             KC_ENT,     KC_SPACE,            UTILITY,         KC_LEFT, KC_DOWN, KC_RGHT
     ),
 
-    [1] = LAYOUT_A_1x4(
+    [1] = LAYOUT (
+             KC_GRV,  KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_F6,   KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,  KC_DEL,
+    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+             _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
+    _______, _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
+    ),
+
+    [2] = LAYOUT (
              _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
     _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
     _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,                   _______, _______,
-    _______, _______, _______, _______,                   _______,                                 _______,     _______, _______, _______
+             _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
+    _______, _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
     ),
 
-    [2] = LAYOUT_A_1x4(
+    [3] = LAYOUT (
              _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
     _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
     _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,                   _______, _______,
-    _______, _______, _______, _______,                   _______,                                 _______,     _______, _______, _______
+             _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
+    _______, _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
+    ),
+
+    [4] = LAYOUT (
+             _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+    _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+             _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
+    _______, _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
     )
 
 };
 
-// layouts
-
-    // [0] = LAYOUT_A_1x3(
-    //          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,                   _______, _______,
-    //          _______, _______, _______,                   _______,                               _______,       _______, _______, _______
-    // )
-
-// Split Spacebar
-
-    // [0] = LAYOUT_B_1x3(
-    //          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,                   _______, _______,
-    //          _______, _______, _______,                _______, _______,                               _______, _______, _______, _______
-    // )
-
-    // [0] = LAYOUT_B_1x4(
-    //          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______,          _______,
-    // _______, _______, _______, _______,                _______, _______,                               _______, _______, _______, _______
-    // )
-
-// Split Shift
-
-    // [0] = LAYOUT_C_1x3(
-    //          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
-    //          _______, _______, _______,                   _______,                                     _______, _______, _______, _______
-    // )
-
-    // [0] = LAYOUT_C_1x3(
-    //          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
-    // _______, _______, _______, _______,                   _______,                                     _______, _______, _______, _______
-    // )
-
-// Split Shift & Split Spacebar
-
-    // [0] = LAYOUT_D_1x3(
-    //          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
-    //          _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
-    // )
-
-    // [0] = LAYOUT_D_1x3(
-    //          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
-    // _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______, _______,
-    // _______, _______, _______, _______,                   _______, _______,                            _______, _______, _______, _______
-    // )
+void encoder_update_user(uint8_t index, bool clockwise) {
+  if (index == 0) {
+    if (clockwise) {
+      tap_code(KC_VOLD);
+    } else {
+      tap_code(KC_VOLU);
+    }
+  }
