+++ keyboards/bang/keymaps/default/macros.c
